requirements.txt
boto3==1.21.44
botocore==1.24.44
jmespath==1.0.0
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
slack-sdk==3.15.2
urllib3==1.26.9
